backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
scikit-learn==1.5.2
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.32.3

